requirements.txt
solana==0.30.2
aiohttp==3.8.5
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
sqlalchemy==2.0.23
pydantic==2.4.2
httpx==0.24.1
python-multipart==0.0.6
asyncio==3.4.3
plotly==5.18.0
dash==2.14.1 